Bug fix (removes duplicate API data)
Previous duplicate-data-removal tools focused on removing instances of duplicate temporal records from non-DSCOVR sources when both ACE and DSCOVR were considered top be the active satellite.,  In data from 2022-03-24T14:32:00 (et al.), the API data have the same satellite name, and are both tagged as the ‘active’ satellite.  This bug-fix allows duplicates of this type to be removed.  I also changed the4 README to correct a small error in the suggested commands for running these programs in Docker.